+++ src/automation_core/__init__.py
@@ -2,6 +2,6 @@ from ._setup import setup
 from .context import Context
 from .errors import ErrorCollector, collect_errors
 
-__version__ = "1.2.2"
+__version__ = "1.3.0"
 
 __all__ = ["setup", "collect_errors", "Context", "ErrorCollector"]